WIP: is_probabilistic_unique refactoring
