+++ python/tests/core/constraints/factories/test_cardinality_metrics.py
@@ -1,4 +1,11 @@
-from whylogs.core.constraints.factories import distinct_number_in_range
+from whylogs.core.constraints.factories import distinct_number_in_range, is_probabilistic_unique
+
+
+def test_probabilistic_unique(builder):
+    builder.add_constraint(is_probabilistic_unique(column_name="animal"))
+    constraint = builder.build()
+    assert not constraint.validate()
+    pass
 
 
 def test_distinct_number_in_range(builder):
